+++ Backend/src/Clases/Utilidades/Tipo.ts
@@ -1,6 +1,6 @@
     export enum Tipo {
         ENTERO,
-        DOUBLE,
+        DECIMAL,
         BOOLEANO,
         CARACTER,
         CADENA,
